Delay plugin renamed files and folder to use CamelCase
