+++ plugins/CMakeLists.txt
@@ -7,7 +7,7 @@ ADD_SUBDIRECTORY(carlabase)
 ADD_SUBDIRECTORY(carlapatchbay)
 ADD_SUBDIRECTORY(carlarack)
 ADD_SUBDIRECTORY(CrossoverEQ)
-ADD_SUBDIRECTORY(delay)
+ADD_SUBDIRECTORY(Delay)
 ADD_SUBDIRECTORY(DualFilter)
 ADD_SUBDIRECTORY(dynamics_processor)
 ADD_SUBDIRECTORY(Eq)
