+++ plugins/Delay/DelayControlsDialog.h
