+++ plugins/Delay/StereoDelay.h
